アニメーションやパーティクルを追加

+++ eziput/Assets/石田オブジェクト/石田が作ったスクリプト/WeaponData.cs
@@ -7,6 +7,6 @@ public class WeaponData : ItemData
 {
     public WeaponType weaponType;
     public int attackBonus;
-    public int varianceBonus;//どのくらいの武器の振れ幅
+    public float varianceBonus;//どのくらいの武器の振れ幅
     public int durability;//上昇する熟練度
 }
